have prettier run on lint (#17)

+++ .eslintrc.js
@@ -82,7 +82,7 @@ module.exports = {
     "prefer-const": 2,
     "prefer-rest-params": 2,
     "prefer-spread": 2,
-    quotes: [2, "single"],
+    quotes: [2, "double"],
     semi: [2, "always"],
     "spaced-comment": [2, "always", { exceptions: ["/"] }],
     "use-isnan": 2,
